+++ src/animation/mod.rs
@@ -6,7 +6,7 @@ pub mod system;
 
 pub use blend_system::BlendTreeSystem;
 pub use blend_tree::{BlendEntry, BlendTree1D};
-pub use player::{AnimationClip, AnimationPlayer, BlendWeight, UvRect};
+pub use player::{AnimationClip, AnimationPlayer, BlendUv, BlendWeight, UvRect};
 pub use state_machine::{
     AnimParam, AnimState, AnimTransition, AnimationStateMachine, StateMachineSystem, TransitionCond,
 };